CMSPlanet.m: removed M_calc prop added rho0 prop (w/ get.()) implemented get/set for rhoi, units based on rho0 added input checks to set.ai so error msgs are easier to follow ConcentricMaclaurinSpheroids.m: casting deltas to double to accommodate indirect set by set.rhoi

+++ ConcentricMaclaurinSpheroids.m
@@ -768,7 +768,7 @@ classdef ConcentricMaclaurinSpheroids < handle
             assert(numel(val) == obj.nlayers,...
                 'length(deltas) = %g ~= nlayers = %g',...
                 numel(val),obj.nlayers) %#ok<MCSUP>
-            obj.deltas = val(:);
+            obj.deltas = double(val(:));
             obj.cooked = false; %#ok<MCSUP>
             obj.fullyCooked = false; %#ok<MCSUP>
